Palworld 0.3.7.58983 Update

+++ Source/Pal/Private/PalExpDatabase.cpp
@@ -5,26 +5,26 @@ UPalExpDatabase::UPalExpDatabase() {
     this->PaldexBonusExpDataTable = NULL;
 }
 
-int32 UPalExpDatabase::GetTotalExp(int32 Level, bool IsPlayer) {
+int64 UPalExpDatabase::GetTotalExp(int32 Level, bool IsPlayer) {
     return 0;
 }
 
-int32 UPalExpDatabase::GetNextExp(int32 Level, bool IsPlayer) {
+int64 UPalExpDatabase::GetNextExp(int32 Level, bool IsPlayer) {
     return 0;
 }
 
-int32 UPalExpDatabase::GetDropExp(int32 Level, FName RowName) {
+int64 UPalExpDatabase::GetDropExp(int32 Level, FName RowName) {
     return 0;
 }
 
-void UPalExpDatabase::DistributionExpValue_forPlayerParty_Server(int32 ExpValue, TArray<APalPlayerCharacter*> GiftPlayerList, bool isCallDelegate) {
+void UPalExpDatabase::DistributionExpValue_forPlayerParty_Server(int64 ExpValue, TArray<APalPlayerCharacter*> GiftPlayerList, bool isCallDelegate) {
 }
 
-int32 UPalExpDatabase::CalcPaldexBonusExp(const FName CharacterID, int32 CaptureCount, APalPlayerState* captureOwnerState) {
+int64 UPalExpDatabase::CalcPaldexBonusExp(const FName CharacterID, int32 CaptureCount, APalPlayerState* captureOwnerState) {
     return 0;
 }
 
-int32 UPalExpDatabase::CalcNeedLevelUpExp(int32 TotalEXP, bool IsPlayer) {
+int64 UPalExpDatabase::CalcNeedLevelUpExp(int32 TotalEXP, bool IsPlayer) {
     return 0;
 }
 
@@ -32,7 +32,7 @@ int32 UPalExpDatabase::CalcLevelFromTotalExp(int32 TotalEXP, bool IsPlayer) {
     return 0;
 }
 
-void UPalExpDatabase::AddExpValue_forPlayerParty_Server(int32 ExpValue, TArray<APalPlayerCharacter*> GiftPlayerList, bool isCallDelegate) {
+void UPalExpDatabase::AddExpValue_forPlayerParty_Server(int64 ExpValue, TArray<APalPlayerCharacter*> GiftPlayerList, bool isCallDelegate) {
 }
 
 void UPalExpDatabase::AddExp_forPlayerParty_TowerBoss(UPalIndividualCharacterHandle* DeadEnemyHandle, TArray<APalPlayerCharacter*> GiftPlayerList, bool isCallDelegate) {
