+++ Source/Pal/Public/PalCaptureBonusExpRaw.h
@@ -8,7 +8,7 @@ struct FPalCaptureBonusExpRaw : public FTableRowBase {
     GENERATED_BODY()
 public:
     UPROPERTY(BlueprintReadWrite, EditAnywhere, meta=(AllowPrivateAccess=true))
-    int32 BonusExp;
+    int64 BonusExp;
     
     PAL_API FPalCaptureBonusExpRaw();
 };
